# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Construction Extracts Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  admin: "admin",
  supervisor: "supervisor",
  contract_supervisor: "contract_supervisor",
  user: "user",
} as const;

export type UserProfileStatus =
  (typeof UserProfileStatus)[keyof typeof UserProfileStatus];

export const UserProfileStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface UserProfile {
  id: number;
  clerkId: string;
  email: string;
  name?: string | null;
  role: UserProfileRole;
  contractCompany?: string | null;
  status: UserProfileStatus;
  company?: string | null;
  phone?: string | null;
  hospital?: string | null;
  jobTitle?: string | null;
  contractNumber?: string | null;
  lastLoginAt?: string | null;
  lastPage?: string | null;
  lastPageAt?: string | null;
  createdAt: string;
}

export interface UserListResponse {
  users: UserProfile[];
  total: number;
  page: number;
  limit: number;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  active: "active",
  completed: "completed",
  on_hold: "on_hold",
} as const;

export interface Project {
  id: number;
  name: string;
  description?: string | null;
  location?: string | null;
  contractValue?: number | null;
  startDate?: string | null;
  endDate?: string | null;
  status: ProjectStatus;
  createdAt: string;
}

export interface CreateProjectBody {
  name: string;
  description?: string | null;
  location?: string | null;
  contractValue?: number | null;
  startDate?: string | null;
  endDate?: string | null;
}

export interface ProjectListResponse {
  projects: Project[];
  total: number;
}

export type ExtractStatus = (typeof ExtractStatus)[keyof typeof ExtractStatus];

export const ExtractStatus = {
  current: "current",
  completed: "completed",
  previous: "previous",
} as const;

export interface Extract {
  id: number;
  extractNumber: string;
  projectId: number;
  projectName?: string | null;
  status: ExtractStatus;
  amount: number;
  description?: string | null;
  submittedBy?: string | null;
  submittedAt?: string | null;
  approvedBy?: string | null;
  approvedAt?: string | null;
  notes?: string | null;
  createdAt: string;
}

export type CreateExtractBodyStatus =
  (typeof CreateExtractBodyStatus)[keyof typeof CreateExtractBodyStatus];

export const CreateExtractBodyStatus = {
  current: "current",
  completed: "completed",
  previous: "previous",
} as const;

export interface CreateExtractBody {
  extractNumber: string;
  projectId: number;
  status: CreateExtractBodyStatus;
  amount: number;
  description?: string | null;
  notes?: string | null;
  submittedAt?: string | null;
}

export type UpdateExtractBodyStatus =
  (typeof UpdateExtractBodyStatus)[keyof typeof UpdateExtractBodyStatus];

export const UpdateExtractBodyStatus = {
  current: "current",
  completed: "completed",
  previous: "previous",
} as const;

export interface UpdateExtractBody {
  status?: UpdateExtractBodyStatus;
  amount?: number;
  description?: string | null;
  notes?: string | null;
  approvedAt?: string | null;
}

export interface ExtractListResponse {
  extracts: Extract[];
  total: number;
  page: number;
  limit: number;
}

export interface DashboardStats {
  totalExtracts: number;
  currentExtracts: number;
  completedExtracts: number;
  previousExtracts: number;
  totalAmount: number;
  currentAmount: number;
  totalProjects: number;
  activeProjects: number;
  pendingUsers: number;
}

export interface ExtractStatusCount {
  status: string;
  count: number;
  amount: number;
}

export interface ExtractActivity {
  id: number;
  extractNumber: string;
  projectName?: string | null;
  status: string;
  amount: number;
  action: string;
  timestamp: string;
}

export type ListUsersParams = {
  status?: ListUsersStatus;
  page?: number;
  limit?: number;
};

export type ListUsersStatus =
  (typeof ListUsersStatus)[keyof typeof ListUsersStatus];

export const ListUsersStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export type ListExtractsParams = {
  status?: ListExtractsStatus;
  projectId?: number;
  page?: number;
  limit?: number;
};

export type ListExtractsStatus =
  (typeof ListExtractsStatus)[keyof typeof ListExtractsStatus];

export const ListExtractsStatus = {
  current: "current",
  completed: "completed",
  previous: "previous",
} as const;