# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Construction Extracts Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      security:
        - clerkAuth: []
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectListResponse"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      security:
        - clerkAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /extracts:
    get:
      operationId: listExtracts
      tags: [extracts]
      summary: List all extracts
      security:
        - clerkAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [current, completed, previous]
        - name: projectId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of extracts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractListResponse"
    post:
      operationId: createExtract
      tags: [extracts]
      summary: Create a new extract
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExtractBody"
      responses:
        "201":
          description: Extract created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Extract"
  /extracts/{extractId}:
    get:
      operationId: getExtract
      tags: [extracts]
      summary: Get extract by ID
      security:
        - clerkAuth: []
      parameters:
        - name: extractId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Extract details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Extract"
    put:
      operationId: updateExtract
      tags: [extracts]
      summary: Update an extract
      security:
        - clerkAuth: []
      parameters:
        - name: extractId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExtractBody"
      responses:
        "200":
          description: Extract updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Extract"
    delete:
      operationId: deleteExtract
      tags: [extracts]
      summary: Delete an extract
      security:
        - clerkAuth: []
      parameters:
        - name: extractId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Extract deleted
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        contractValue:
          type: number
          nullable: true
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        status:
          type: string
          enum: [active, completed, on_hold]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - status
        - createdAt
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        contractValue:
          type: number
          nullable: true
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
      required:
        - name
    ProjectListResponse:
      type: object
      properties:
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        total:
          type: integer
      required:
        - projects
        - total
    Extract:
      type: object
      properties:
        id:
          type: integer
        extractNumber:
          type: string
        projectId:
          type: integer
        projectName:
          type: string
          nullable: true
        status:
          type: string
          enum: [current, completed, previous]
        amount:
          type: number
        description:
          type: string
          nullable: true
        submittedBy:
          type: string
          nullable: true
        submittedAt:
          type: string
          format: date-time
          nullable: true
        approvedBy:
          type: string
          nullable: true
        approvedAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - extractNumber
        - projectId
        - status
        - amount
        - createdAt
    CreateExtractBody:
      type: object
      properties:
        extractNumber:
          type: string
        projectId:
          type: integer
        status:
          type: string
          enum: [current, completed, previous]
        amount:
          type: number
        description:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        submittedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - extractNumber
        - projectId
        - status
        - amount
    UpdateExtractBody:
      type: object
      properties:
        status:
          type: string
          enum: [current, completed, previous]
        amount:
          type: number
        description:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        approvedAt:
          type: string
          format: date-time
          nullable: true
    ExtractListResponse:
      type: object
      properties:
        extracts:
          type: array
          items:
            $ref: "#/components/schemas/Extract"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - extracts
        - total
        - page
        - limit
...