# lib/api-zod/src/generated/types/extractListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Construction Extracts Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { Extract } from "./extract";

export interface ExtractListResponse {
  extracts: Extract[];
  total: number;
  page: number;
  limit: number;
}