# lib/api-zod/src/generated/types/updateExtractBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Construction Extracts Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateExtractBodyStatus } from "./updateExtractBodyStatus";

export interface UpdateExtractBody {
  status?: UpdateExtractBodyStatus;
  amount?: number;
  description?: string | null;
  notes?: string | null;
  approvedAt?: Date | null;
}